Version: Fix RELEASE_TARBALL case in mixed chroot environment
Uncovered by haikuports, because that environment uses a git
directory during the build control process, but the chroot it drops
into for the build does not have git available.  This resulted in
the RELEASE_TARBALL check failing because the `.git` directory
had reappeared, but then git wasn't there so Version.cmake failed.