ImageSeq: rely on CMake-internal find_package on all platforms
CMake's find_package was already used on everything other than
Windows, but as it turns out, on Windows, the user only needs
to pass -DCMAKE_PREFIX_PATH with the root directory of where
the DevIL installation is for it to find it.

The DevIL SDK as distributed from upstream requires either:
A) Moving the relevant x64 or x86 version of DevIL.lib and ILU.lib
   into the main lib/ directory instead of them residing in
   lib/{x64,x86}/Release or lib/{x64,x86}/unicode/Release
or
B) Using -DIL_LIBRARIES and -DILU_LIBRARIES when configuring
   AviSynth+ to point directly at DevIL.lib and ILU.lib.

This also allows for linking against a static build of DevIL,
removing the need for any extra system DLLs.