Make configurable user plugindir more robust
There was no need for it to use ENV${HOME} in CMakeLists.txt,
when the usage of $HOME/.avisynth in the autoload defaults already
used a more robust tactic to find the value of $HOME. Switch
USER_AVS_PLUGINDIR_LOCATION to the same method.

It still defaults to .local/lib/avisynth.