buildsystem: restrict use of MSVC to x86(-64) only
MSVC (including ClangCL) and GCC (including standard Clang) are
not C++ compatible, and this has caused no end of headaches on
ostensibly 'simple' matters of compiling plugins for x86(-64).

Windows on ARM is still a relatively niche target, there has not
yet been a truly fleshed-out release of AviSynth+ for it, so let's
avoid repeating this mistake.  There is no legacy cache of native
ARM plugins to get messed up by this, and there is the upside of
any future architecture choices not being hampered by requiring
any compatibility with MSVC.

Windows on ARM builds are therefore to be built with MinGW, and
the MSVC codepaths are fused off for any non-x86(-64) arch.