docs: replace mentions of avisynth.org with avisynth.nl
avisynth.org has been defunct/not in the control of any of the
old devs for a while now, and worse, it's now been domain squatted
by spammers.

Where possible, directly change the documentation to the current
avisynth.nl address for the same page.  If the page no longer
exists, either pull a working archive.org version of the page,
or remove it from the docs or change it to a more appropriate
link (for instance, the known issues pages for 2.52/2.53).